Code and GUI updated

+++ RackSimSoft/src/Gui2D/RectangleComponent.java
@@ -13,18 +13,16 @@ public class RectangleComponent extends JComponent
 	
 	int width = 0;
 	int heigth = 0;
-	int border = 0;
+	float border = 0.0f;
 	String id = "";
 	
-	public RectangleComponent(String id, int width, int heigth, int border)
+	public RectangleComponent(String id, int width, int heigth, float border)
 	{
 		this.width = width;
 		this.heigth = heigth;
 		this.border = border;
 		
 		this.id = id;
-		
-		//System.out.println("RectangleComponent " + this.id + " wurde erzeugt...");
 	}
 	
 	public void paintComponent(Graphics g)
@@ -38,7 +36,5 @@ public class RectangleComponent extends JComponent
 		
 		g2.draw(rectangle);
 		g2.drawString("" + this.id, (this.width / 2), (this.heigth / 2));
-		
-		//System.out.println("RectangleComponent " + this.id + " wurde gezeichnet...");
 	}
 }
